master: flutter demo app v1.4.0

+++ packages/clean_code_infra/lib/src/database/hive_local_database.dart
@@ -18,10 +18,12 @@ class HiveLocalDatabase<T>
         _dbName = databaseName;
         await Hive.initFlutter();
         _hiveBox = await Hive.openBox<String>(databaseName);
-        Log.success(
-          'Local Database [$databaseName] init successful',
-          throwsCrashlytics: false,
-        );
+        if (!Log.isIntegrationTest) {
+          Log.success(
+            'Local Database [$databaseName] init successful',
+            throwsCrashlytics: false,
+          );
+        }
       } catch (error, stackTrace) {
         _throw('init', error, stackTrace);
       }
@@ -31,7 +33,9 @@ class HiveLocalDatabase<T>
   @override
   Future<void> close() async {
     try {
-      Log.info('[$_dbName].close');
+      if (!Log.isIntegrationTest) {
+        Log.info('[$_dbName].close');
+      }
       return await _hiveBox.close();
     } catch (error, stackTrace) {
       _throw('close', error, stackTrace);
@@ -41,7 +45,9 @@ class HiveLocalDatabase<T>
   @override
   Future<int> add(T value) async {
     try {
-      Log.info('[$_dbName].add \n$value');
+      if (!Log.isIntegrationTest) {
+        Log.info('[$_dbName].add \n$value');
+      }
       return await _hiveBox.add(encodeValue<T>(value));
     } catch (error, stackTrace) {
       _throw('add', error, stackTrace);
@@ -51,7 +57,9 @@ class HiveLocalDatabase<T>
   @override
   Future<void> addByKey(String key, T value) async {
     try {
-      Log.info('[$_dbName].addByKey [$key] \n$value');
+      if (!Log.isIntegrationTest) {
+        Log.info('[$_dbName].addByKey [$key] \n$value');
+      }
       return await _hiveBox.put(key, encodeValue<T>(value));
     } catch (error, stackTrace) {
       _throw('addByKey', error, stackTrace);
@@ -63,13 +71,17 @@ class HiveLocalDatabase<T>
     try {
       final String? value = _hiveBox.get(key);
       if (value == null) {
-        Log.warning(
-          '[$_dbName].get [$key] not found',
-          throwsCrashlytics: false,
-        );
+        if (!Log.isIntegrationTest) {
+          Log.warning(
+            '[$_dbName].get [$key] not found',
+            throwsCrashlytics: false,
+          );
+        }
         return null;
       }
-      Log.success('[$_dbName].get [$key]', throwsCrashlytics: false);
+      if (!Log.isIntegrationTest) {
+        Log.success('[$_dbName].get [$key]', throwsCrashlytics: false);
+      }
       return decodeValue<T>(value);
     } catch (error, stackTrace) {
       _throw('get', error, stackTrace);
@@ -79,7 +91,9 @@ class HiveLocalDatabase<T>
   @override
   Future<void> update(dynamic key, T value) async {
     try {
-      Log.info('[$_dbName].update [$key] \n$value');
+      if (!Log.isIntegrationTest) {
+        Log.info('[$_dbName].update [$key] \n$value');
+      }
       return await _hiveBox.put(key, encodeValue<T>(value));
     } catch (error, stackTrace) {
       _throw('update', error, stackTrace);
@@ -90,13 +104,17 @@ class HiveLocalDatabase<T>
   Future<void> delete(dynamic key) async {
     try {
       if (_hiveBox.get(key) == null) {
-        Log.warning(
-          '[$_dbName].delete [$key] not found',
-          throwsCrashlytics: false,
-        );
+        if (!Log.isIntegrationTest) {
+          Log.warning(
+            '[$_dbName].delete [$key] not found',
+            throwsCrashlytics: false,
+          );
+        }
         throw NotFoundException();
       } else {
-        Log.success('[$_dbName].delete [$key]', throwsCrashlytics: false);
+        if (!Log.isIntegrationTest) {
+          Log.success('[$_dbName].delete [$key]', throwsCrashlytics: false);
+        }
         return await _hiveBox.delete(key);
       }
     } catch (error, stackTrace) {
@@ -109,10 +127,14 @@ class HiveLocalDatabase<T>
     try {
       if (keys == null) {
         await _hiveBox.clear();
-        Log.info('[$_dbName].deleteAll');
+        if (!Log.isIntegrationTest) {
+          Log.info('[$_dbName].deleteAll');
+        }
         return;
       }
-      Log.info('[$_dbName].deleteAll keys $keys');
+      if (!Log.isIntegrationTest) {
+        Log.info('[$_dbName].deleteAll keys $keys');
+      }
       return await _hiveBox.deleteAll(keys);
     } catch (error, stackTrace) {
       _throw('deleteAll', error, stackTrace);
@@ -125,7 +147,9 @@ class HiveLocalDatabase<T>
       final values = _hiveBox.toMap().map<dynamic, T>((key, value) {
         return MapEntry(key, decodeValue<T>(value));
       });
-      Log.info('[$_dbName].values [${values.length}] length');
+      if (!Log.isIntegrationTest) {
+        Log.info('[$_dbName].values [${values.length}] length');
+      }
       return values;
     } catch (error, stackTrace) {
       _throw('values', error, stackTrace);
